+++ src/etoolbox/datazip/wrapper.py
@@ -1,6 +1,7 @@
 """Use of :class:`.IOMixin` as basis for a wrapper."""
 from __future__ import annotations
 
+import logging
 from datetime import datetime
 from io import BytesIO
 from pathlib import Path
@@ -8,8 +9,9 @@ from typing import Any
 from zipfile import ZIP_STORED
 from zoneinfo import ZoneInfo
 
-from etoolbox import __version__
 from etoolbox.datazip.mixin import IOMixin
+
+LOGGER = logging.getLogger(__name__)
 
 
 class IOWrapper(IOMixin):
@@ -26,13 +28,13 @@ class IOWrapper(IOMixin):
                 organized like :py:const:`etoolbox.datazip.core.RECIPES`
         """
         self._obj = obj
-        self._metadata = {
+        self._metadata: dict = {
             "created": str(datetime.now(tz=ZoneInfo("UTC"))),
-            "version": __version__,
         }
-        self._recipes = {} if recipes is None else recipes
+        self._recipes: dict = {} if recipes is None else recipes
 
-    def __getattr__(self, item):
+    def __getattr__(self, item) -> Any:
+        """Pretend to be `self._obj`."""
         return getattr(self._obj, item)
 
     def to_file(
@@ -41,7 +43,7 @@ class IOWrapper(IOMixin):
         compression=ZIP_STORED,
         clobber=False,
         **kwargs,
-    ):
+    ) -> None:
         """Write out the obj to a file."""
         self._to_file(
             self._obj,
@@ -54,13 +56,12 @@ class IOWrapper(IOMixin):
         )
 
     @classmethod
-    def from_file(cls, path, **kwargs):
+    def from_file(cls, path: Path | str | BytesIO, **kwargs) -> Any:
         """Recreate an instance of :class:`.DataZipWrapper` with the wrapped object."""
-        obj, metadata = cls._from_file(None, path, cls_from_meta=True, **kwargs)
+        obj, metadata = cls._from_file(None, path, **kwargs)
         self = cls(obj)
         self._metadata = metadata
         return self
 
-    def __repr__(self):
-        qname = self._obj.__class__.__qualname__
-        return f"IOWrap{qname}({repr(self._obj).removeprefix(qname + '(').removesuffix(')')})"
+    def __repr__(self) -> str:
+        return self.__class__.__qualname__ + f"({repr(self._obj)})"